// src/modules/preventive-maintenances/models/task-categories.enum.ts
export enum TaskCategories {
  CORRECTIVE_MAINTENANCE = 'CORRECTIVE_MAINTENANCE',
  CLEANUP = 'CLEANUP',
  REPLACEMENT = 'REPLACEMENT',
  OTHERS = 'OTHERS',
  PREVENTIVE_MAINTENANCE = 'PREVENTIVE_MAINTENANCE',
  DAMAGE = 'DAMAGE',
  INSPECTION = 'INSPECTION',
  SAFETY = 'SAFETY',
  TASK = 'TASK',
  DEFICIENCY = 'DEFICIENCY',
}